// src/openapi/model/pROCESSINGTYPE.ts
/**
 * ICNC.ERP.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type PROCESSINGTYPE = 'raw' | 'blacken' | 'bonderizing' | 'nickelPlate' | 'heatTreatment' | 'wedM_LS' | 'wedM_HS' | 'matteTin' | 'goldPlating' | 'silverPlating' | 'electrolessNickelPlating' | 'trivalentColorZincPlating' | 'turning' | 'chamfer' | 'grind' | 'milling' | 'spraying' | 'blunt' | 'zincNickelAlloyPlating' | 'hexavalentColorZincPlating' | 'threadForming' | 'qpq' | 'yellowZincPlating' | 'whiteZincPlating' | 'blackZincPlating' | 'brightZincPlating' | 'assemble' | 'dispensing';

export const PROCESSINGTYPE = {
    Raw: 'raw' as PROCESSINGTYPE,
    Blacken: 'blacken' as PROCESSINGTYPE,
    Bonderizing: 'bonderizing' as PROCESSINGTYPE,
    NickelPlate: 'nickelPlate' as PROCESSINGTYPE,
    HeatTreatment: 'heatTreatment' as PROCESSINGTYPE,
    WedMLs: 'wedM_LS' as PROCESSINGTYPE,
    WedMHs: 'wedM_HS' as PROCESSINGTYPE,
    MatteTin: 'matteTin' as PROCESSINGTYPE,
    GoldPlating: 'goldPlating' as PROCESSINGTYPE,
    SilverPlating: 'silverPlating' as PROCESSINGTYPE,
    ElectrolessNickelPlating: 'electrolessNickelPlating' as PROCESSINGTYPE,
    TrivalentColorZincPlating: 'trivalentColorZincPlating' as PROCESSINGTYPE,
    Turning: 'turning' as PROCESSINGTYPE,
    Chamfer: 'chamfer' as PROCESSINGTYPE,
    Grind: 'grind' as PROCESSINGTYPE,
    Milling: 'milling' as PROCESSINGTYPE,
    Spraying: 'spraying' as PROCESSINGTYPE,
    Blunt: 'blunt' as PROCESSINGTYPE,
    ZincNickelAlloyPlating: 'zincNickelAlloyPlating' as PROCESSINGTYPE,
    HexavalentColorZincPlating: 'hexavalentColorZincPlating' as PROCESSINGTYPE,
    ThreadForming: 'threadForming' as PROCESSINGTYPE,
    Qpq: 'qpq' as PROCESSINGTYPE,
    YellowZincPlating: 'yellowZincPlating' as PROCESSINGTYPE,
    WhiteZincPlating: 'whiteZincPlating' as PROCESSINGTYPE,
    BlackZincPlating: 'blackZincPlating' as PROCESSINGTYPE,
    BrightZincPlating: 'brightZincPlating' as PROCESSINGTYPE,
    Assemble: 'assemble' as PROCESSINGTYPE,
    Dispensing: 'dispensing' as PROCESSINGTYPE
};